requirements/base.txt
black==22.12.0
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
isort==5.11.4
python-decouple==3.6
psycopg2==2.9.5
Pillow
celery
django-cors-headers==3.13.0
drf-yasg==1.21.4
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
Faker
pytest-django
pytest-factoryboy
redis
django-celery-beat
django-celery-results
drf-nested-routers


